+++ xclim/sdba/base.py
@@ -571,6 +571,7 @@ def map_blocks(reduces=None, **outvars):
             # All dimensions of the output data, new_dims are added at the end on purpose.
             all_dims = base_dims + new_dims
             # The coordinates of the output data.
+            added_coords = []
             coords = {}
             sizes = {}
             for dim in all_dims:
@@ -578,13 +579,15 @@ def map_blocks(reduces=None, **outvars):
                     coords[group.prop] = group.get_coordinate(ds=ds)
                 elif dim == group.dim:
                     coords[group.dim] = ds[group.dim]
-                elif dim in ds.coords:
-                    coords[dim] = ds[dim]
                 elif dim in kwargs:
                     coords[dim] = xr.DataArray(kwargs[dim], dims=(dim,), name=dim)
                 elif dim in ds.dims:
-                    # Dimension with no coordinate, but it is in the inputs so we can get the size.
-                    sizes[dim] = ds[dim].size
+                    # If a dim has no coords : some sdba function will add them, so to be safe we add them right now
+                    # and note them to remove them afterwards.
+                    if dim not in ds.coords:
+                        added_coords.append(dim)
+                    ds[dim] = ds[dim]
+                    coords[dim] = ds[dim]
                 else:
                     raise ValueError(
                         f"This function adds the {dim} dimension, its coordinate must be provided as a keyword argument."
@@ -626,18 +629,34 @@ def map_blocks(reduces=None, **outvars):
                     raise ValueError(
                         f"{func.__name__} failed on block with coords : {dsblock.coords}."
                     ) from err
-                for name, crd in dsblock.coords.items():
-                    if name not in out.coords and set(crd.dims).issubset(out.dims):
-                        out = out.assign_coords({name: dsblock[name]})
                 return out
 
             # Fancy patching for explicit dask task names
             _call_and_transpose_on_exit.__name__ = f"block_{func.__name__}"
 
+            # Remove all auxiliary coords on both tmpl and ds
+            extra_coords = {
+                nam: crd for nam, crd in ds.coords.items() if nam not in crd.dims
+            }
+            ds = ds.drop_vars(extra_coords.keys())
+            tmpl = tmpl.drop_vars(extra_coords.keys())
+
+            # Call
             out = ds.map_blocks(
                 _call_and_transpose_on_exit, template=tmpl, kwargs=kwargs
             )
 
+            # Add back the extra coords, but only those which have compatible dimensions (like xarray would have done)
+            out = out.assign_coords(
+                {
+                    nam: crd
+                    for nam, crd in extra_coords.items()
+                    if set(crd.dims).issubset(out.dims)
+                }
+            )
+
+            # Finally remove coords we added... 'ignore' in case they were already removed.
+            out = out.drop_vars(added_coords, errors="ignore")
             return out
 
         _map_blocks.__dict__["func"] = func