+++ xclim/testing/tests/test_precip.py
@@ -5,7 +5,7 @@ import pandas as pd
 import pytest
 import xarray as xr
 
-from xclim import atmos, set_options
+from xclim import atmos, core, set_options
 from xclim.core.calendar import percentile_doy
 from xclim.testing import open_dataset
 
@@ -125,6 +125,21 @@ class TestWetDays:
         # make sure that vector with all nans gives nans whatever skipna
         assert np.isnan(out1.values[0, -1, -1])
         # assert (np.isnan(wds.values[0, -1, -1]))
+
+
+class TestWetPrcptot:
+    """Testing of prcptot with wet days"""
+
+    def test_simple(self):
+        pr = open_dataset("ERA5/daily_surface_cancities_1990-1993.nc").pr
+
+        thresh = "1 mm/day"
+        out = atmos.wet_precip_accumulation(pr, thresh=thresh)
+
+        # Reference value
+        t = core.units.convert_units_to(thresh, pr)
+        pa = atmos.precip_accumulation(pr.where(pr >= t, 0))
+        np.testing.assert_array_equal(out, pa)
 
 
 class TestDailyIntensity:
@@ -383,10 +398,10 @@ def test_days_over_precip_thresh():
     per = percentile_doy(pr, window=5, per=80)
 
     out1 = atmos.days_over_precip_thresh(pr, per)
-    np.testing.assert_array_equal(out1[1, :, 0], np.array([87, 63, 74, 79]))
+    np.testing.assert_array_equal(out1[1, :, 0], np.array([81, 61, 69, 78]))
 
     out2 = atmos.days_over_precip_thresh(pr, per, thresh="2 mm/d")
-    np.testing.assert_array_equal(out2[1, :, 0], np.array([87, 63, 71, 79]))
+    np.testing.assert_array_equal(out2[1, :, 0], np.array([81, 61, 66, 78]))
 
     assert "only days with at least 2 mm/d are counted." in out2.description
 
@@ -397,12 +412,12 @@ def test_fraction_over_precip_thresh():
 
     out = atmos.fraction_over_precip_thresh(pr, per)
     np.testing.assert_allclose(
-        out[1, :, 0], np.array([0.832, 0.785, 0.776, 0.815]), atol=0.001
+        out[1, :, 0], np.array([0.809, 0.770, 0.748, 0.807]), atol=0.001
     )
 
     out = atmos.fraction_over_precip_thresh(pr, per, thresh="0.002 m/d")
     np.testing.assert_allclose(
-        out[1, :, 0], np.array([0.853, 0.818, 0.803, 0.840]), atol=0.001
+        out[1, :, 0], np.array([0.831, 0.803, 0.774, 0.833]), atol=0.001
     )
 
     assert "only days with at least 0.002 m/d are included" in out.description